refactor(monolith): fold chunker into knowledge and k8s client into home/observability
