+++ projects/monolith/home/observability/kubernetes_test.py
@@ -7,7 +7,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from shared.kubernetes import KubernetesClient, _parse_cpu, _parse_memory
+from home.observability.kubernetes import KubernetesClient, _parse_cpu, _parse_memory
 
 
 @pytest.fixture
@@ -39,9 +39,9 @@ async def test_count_nodes(k8s_client):
     )
 
     with (
-        patch("shared.kubernetes.config.load_incluster_config"),
-        patch("shared.kubernetes.ApiClient", return_value=mock_api),
-        patch("shared.kubernetes.client.CoreV1Api", return_value=mock_v1),
+        patch("home.observability.kubernetes.config.load_incluster_config"),
+        patch("home.observability.kubernetes.ApiClient", return_value=mock_api),
+        patch("home.observability.kubernetes.client.CoreV1Api", return_value=mock_v1),
     ):
         count = await k8s_client.count_nodes()
 
@@ -59,9 +59,12 @@ async def test_count_argocd_applications(k8s_client):
     )
 
     with (
-        patch("shared.kubernetes.config.load_incluster_config"),
-        patch("shared.kubernetes.ApiClient", return_value=mock_api),
-        patch("shared.kubernetes.client.CustomObjectsApi", return_value=mock_custom),
+        patch("home.observability.kubernetes.config.load_incluster_config"),
+        patch("home.observability.kubernetes.ApiClient", return_value=mock_api),
+        patch(
+            "home.observability.kubernetes.client.CustomObjectsApi",
+            return_value=mock_custom,
+        ),
     ):
         count = await k8s_client.count_argocd_applications()
 
@@ -115,10 +118,13 @@ async def test_aggregate_node_resources_uses_rss_not_working_set(k8s_client):
     )
 
     with (
-        patch("shared.kubernetes.config.load_incluster_config"),
-        patch("shared.kubernetes.ApiClient", return_value=mock_api),
-        patch("shared.kubernetes.client.CoreV1Api", return_value=mock_v1),
-        patch("shared.kubernetes.client.CustomObjectsApi", return_value=mock_custom),
+        patch("home.observability.kubernetes.config.load_incluster_config"),
+        patch("home.observability.kubernetes.ApiClient", return_value=mock_api),
+        patch("home.observability.kubernetes.client.CoreV1Api", return_value=mock_v1),
+        patch(
+            "home.observability.kubernetes.client.CustomObjectsApi",
+            return_value=mock_custom,
+        ),
     ):
         result = await k8s_client.aggregate_node_resources()
 
@@ -183,8 +189,8 @@ async def test_close_cleans_up(k8s_client):
     mock_api.close = AsyncMock()
 
     with (
-        patch("shared.kubernetes.config.load_incluster_config"),
-        patch("shared.kubernetes.ApiClient", return_value=mock_api),
+        patch("home.observability.kubernetes.config.load_incluster_config"),
+        patch("home.observability.kubernetes.ApiClient", return_value=mock_api),
     ):
         # Force client creation
         await k8s_client._ensure_client()
